extract model, rename cloudformed entities
