+++ cfn.yaml
@@ -70,10 +70,10 @@ Resources:
                 - dynamodb:Query
               Resource: !Sub arn:aws:dynamodb:${AWS::Region}:${AWS::AccountId}:table/${App}-${Stage}-articles
 
-  SavedArticleLambdaPermission:
+  FetchArticlesLambdaPermission:
     Type: AWS::Lambda::Permission
     Properties:
-      FunctionName: !GetAtt SavedArticlesLambda.Arn
+      FunctionName: !GetAtt FetchArticlesLambda.Arn
       Action: lambda:InvokeFunction
       Principal: apigateway.amazonaws.com
 
@@ -87,23 +87,25 @@ Resources:
           version: "1.0.0"
           title: !Ref App
         paths:
-          "/syncedPrefs/me/savedArticles":
+
+          "/syncedPrefs/me":
             get:
               x-amazon-apigateway-integration:
                 httpMethod: POST
                 type: aws_proxy
-                uri: !Sub arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SavedArticlesLambda.Arn}/invocations
+                uri: !Sub arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${FetchArticlesLambda.Arn}/invocations
               produces:
               - application/json
               responses:
                 "200":
                   "description": "200 response"
 
+          "/syncedPrefs/me/savedArticles":
             post:
               x-amazon-apigateway-integration:
                 httpMethod: POST
                 type: aws_proxy
-                uri: !Sub arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SaveForLaterLambda.Arn}/invocations
+                uri: !Sub arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SaveArticlesLambda.Arn}/invocations
               consumes:
               - application/json
               produces:
@@ -140,10 +142,10 @@ Resources:
       Stage: !Ref Stage
 
 
-  SaveForLaterLambda:
+  SaveArticlesLambda:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: com.gu.sfl.lambda.SaveForLaterLambda::handleRequest
+      Handler: com.gu.sfl.lambda.SaveArticlesLambda::handleRequest
       Role: !GetAtt SaveForLaterRole.Arn
       Runtime: java8
       CodeUri:
@@ -165,16 +167,16 @@ Resources:
            Method: POST
            RestApiId: !Ref SaveForLaterApi
 
-  SavedArticlesLambda:
+  FetchArticlesLambda:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: com.gu.sfl.lambda.SavedArticlesLambda::handleRequest
+      Handler: com.gu.sfl.lambda.FetchArticlesLambda::handleRequest
       Role: !GetAtt SaveForLaterRole.Arn
       Runtime: java8
       CodeUri:
         Bucket: !Ref DeployBucket
         Key: !Sub ${Stack}/${Stage}/${App}/${App}.jar
-      FunctionName: !Sub ${App}-SAVED-${Stage}
+      FunctionName: !Sub ${App}-FETCH-${Stage}
       Environment:
         Variables:
           Stack: !Ref Stack
@@ -186,7 +188,7 @@ Resources:
         GetApi:
           Type: Api
           Properties:
-           Path: "/syncedPrefs/me/SavedArticles"
+           Path: "/syncedPrefs/me"
            Method: GET
            RestApiId: !Ref SaveForLaterApi
