+++ src/main/scala/com/gu/sfl/controller/SavedArticlesController.scala
@@ -1,17 +1,10 @@
 package com.gu.sfl.controller
 
-import java.util.concurrent.TimeUnit
-
-import com.gu.sfl.identity.IdentityService
-import com.gu.sfl.{Logging, Parallelism}
 import com.gu.sfl.lambda.{LambdaRequest, LambdaResponse}
-import com.gu.sfl.persisitence.SavedArticlesPersistence
-import com.gu.sfl.util.StatusCodes
-import com.gu.sfl.lib.Jackson.mapper
 import com.gu.sfl.savedarticles.FetchSavedArticles
+import com.gu.sfl.{Logging, Parallelism}
 
-import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
 
 
@@ -25,7 +18,9 @@ class SavedArticlesController(fetchSavedArticles: FetchSavedArticles) extends Fu
 
    val futureResponse =  fetchSavedArticles.retrieveSavedArticlesForUser(lambdaRequest.headers).transformWith {
      case Success(Some(syncedPrefs)) =>
-       logger.info(s"Returning found ${syncedPrefs.size} articles")
+       syncedPrefs.savedArticles.map( sa =>
+          logger.info(s"Returning found ${sa.articles.size} articles")
+       )
        Future { okSavedArticlesResponse(syncedPrefs) }
      case Success(None) =>
        logger.info("No articles found")
