+++ src/main/scala/com/gu/sfl/lambda/AwsLambda.scala
@@ -3,11 +3,7 @@ package com.gu.sfl.lambda
 import java.io.{InputStream, OutputStream}
 
 import com.amazonaws.services.lambda.runtime.{Context, RequestStreamHandler}
-import com.gu.AwsIdentity
 import com.gu.sfl.Logging
-import com.gu.sfl.lib.SsmConfig
-import com.gu.sfl.controller.SaveForLaterControllerImpl
-import com.gu.sfl.persisitence.{PersistanceConfig, SavedArticlesPersistence, SavedArticlesPersistenceImpl}
 
 import scala.concurrent.Future
 import scala.util.Try
