+++ src/main/scala/com/gu/sfl/lambda/SavedArticlesLambda.scala
@@ -7,7 +7,6 @@ import com.gu.sfl.identity.{IdentityConfig, IdentityServiceImpl}
 import com.gu.sfl.lib.{GlobalHttpClient, SsmConfig}
 import com.gu.sfl.persisitence.{PersistanceConfig, SavedArticlesPersistenceImpl}
 import com.gu.sfl.savedarticles.FetchSavedArticlesImpl
-import com.gu.sfl.identity.IdentityServiceImpl
 
 //TODO - BOTH lambdas need renaming
 object SavedArticlesLambda extends Logging {
