+++ src/main/scala/com/gu/sfl/lib/AuthHeaderParser.scala
@@ -5,8 +5,7 @@ import com.gu.sfl.util.HeaderNames.Identity
 
 trait AuthHeaderParser {
     def getIdentityHeaders(headers: Map[String, String]) : Option[IdentityHeaders] = {
-
-      //IOS sends 'authorisation' whereas android 'Authorisation' 
+      //Ios sends 'authorisation' whereas android 'Authorisation'
       val lowerHeaders = headers.map { case(key, value) => (key.toLowerCase, value)}
       for {
         auth <- lowerHeaders.get(Identity.auth)
