+++ src/test/scala/com/gu/sfl/lib/ArticleMergeSpecification.scala
@@ -2,11 +2,11 @@ package com.gu.sfl.lib
 
 import java.time.LocalDateTime
 
-import com.gu.sfl.controller.{SavedArticle, SavedArticles, SyncedPrefs}
 import com.gu.sfl.exception.{MaxSavedArticleTransgressionError, SavedArticleMergeError}
-import com.gu.sfl.persisitence.{SavedArticlesPersistence, SavedArticlesPersistenceImpl}
-import org.specs2.mutable.Specification
+import com.gu.sfl.model.{SavedArticle, SavedArticles, SyncedPrefs}
+import com.gu.sfl.persisitence.SavedArticlesPersistenceImpl
 import org.specs2.mock.Mockito
+import org.specs2.mutable.Specification
 import org.specs2.specification.Scope
 
 import scala.util.{Failure, Success}
